requirements.txt
openai
streamlit
pandas
langchain
langchain-community
langchain-openai
unstructured
Pillow
google-generativeai
backoff
